+++ src/App.css
@@ -4,8 +4,12 @@
   box-sizing: border-box;
 }
 
-.App {
+.App h1, .card-title, .card-body {
   text-align: center;
+}
+
+.card-body .card-text {
+  text-align: start;
 }
 
 .search-bar-container {
