--- src/App/store.js
@@ -1,5 +1,0 @@
-//import { configureStore } from "@reduxjs/toolkit";
-//
-//export const store = configureStore({
-//	reducer: {},
-//});
