+++ src/Components/Component/search.jsx
@@ -7,7 +7,18 @@ import Form from 'react-bootstrap/Form';
 import InputGroup from 'react-bootstrap/InputGroup';
 
 function SearchBar() {
-	return 
+	const [search, setSearch] = useState("");
+
+	const searBarInput = 
+
+	return (
+		<Form className="input-wrapper d-flex flex-column p-0">
+			<InputGroup className="input-wrapper w-100 align-items-center">
+				<FaSearch id="search-icon" />
+				<Form.Control onChange={(e) => setSearch(e.target.value)} placeholder="Search contacts" className="border-0" id="moreInputStyle" />
+			</InputGroup>
+		</Form>
+	)
 }
 
 export default SearchBar;
