+++ src/Components/Layouts/ContactPage.js
@@ -1,13 +1,11 @@
 import React from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Card from "react-bootstrap/Card";
-import Image from "react-bootstrap/Image";
 
-function ContactUsLayout() {
+function ContactUs() {
 	return (
 		<div className="pt-3 w-75 m-auto pb-5">
 			<h1>Contact us!</h1>
 		</div>
 	);
 }
-export default ContactUsLayout;
+export default ContactUs;
