+++ src/Components/Layouts/GlobalLayout.js
@@ -2,13 +2,15 @@ import CollapsibleNavbar from "../Component/Navbar.js";
 import "bootstrap/dist/css/bootstrap.min.css";
 import Footer from "../Component/Footer.js";
 import "./GlobalStyle.css";
+import { Outlet } from "react-router-dom";
 
 function GlobalLayout() {
 	return (
-		<div className="App">
+		<>
 			<CollapsibleNavbar />
+			<Outlet />
 			<Footer />
-		</div>
+		</>
 	);
 }
