+++ src/Components/Layouts/NotFoundPage.css
@@ -1,0 +1,4 @@
+img {
+    max-width: 500px;
+    width: 100%
+}
