+++ src/Components/Layouts/NotFoundPage.js
@@ -1,0 +1,12 @@
+import React from "react";
+import "./NotFoundPage.css";
+
+function RouteNotFound() {
+	return (
+		<div className=" m-auto mt-5">
+			<img src="ORFI0N0.jpg" alt="Error! Page not found." />
+		</div>
+	);
+}
+
+export default RouteNotFound;
