+++ src/Components/Layouts/ProductPage.css
@@ -1,0 +1,18 @@
+.Product .card{
+    max-width: 50rem; 
+    box-Shadow: 5px 5px 5px rgba(155, 155, 155, 0.50);
+    margin: 20px auto 100px auto;
+}
+
+.Product img {
+    margin: auto;
+}
+
+.Product .card-title {
+    margin-top: 20px;
+}
+
+.product .card-footer, .product .card-text {
+    display: flex;
+    justify-content: flex-start;
+}
